Removed the non event chart components
