+++ demo/Chartist.Blazor.Demo.WASM/Pages/EventedBar.razor
@@ -29,7 +29,7 @@
         <button class="btn btn-primary" @onclick="@ClearSelection">Clear Selection</button>
     </div>
     <div class="col-8">
-        <ChartistEventedBar Data="@chartData" Options="@chartOptions" OnDataPointClicked="DataPointClicked" />
+        <ChartistBar Data="@chartData" Options="@chartOptions" OnDataPointClicked="DataPointClicked" />
     </div>
 </div>
