+++ demo/Chartist.Blazor.Demo.WASM/Pages/EventedLine.razor
@@ -30,7 +30,7 @@
         <button class="btn btn-primary" @onclick="@ClearSelection">Clear Selection</button>
     </div>
     <div class="col-8">
-        <ChartistEventedLine Data="@chartData" Options="@chartOptions" OnDataPointClicked="DataPointClicked" />
+        <ChartistLine Data="@chartData" Options="@chartOptions" OnDataPointClicked="DataPointClicked" />
     </div>
 </div>
