+++ demo/Chartist.Blazor.Demo.WASM/Pages/SimpleLine.razor
@@ -1,6 +1,5 @@
 ﻿@page "/SimpleLine"
 
-
 <h3>Simple Line</h3>
 
 <ChartistLine Data="SimpleLineData" Options="DefaultLineOptions" />
