+++ src/Chartist.Blazor/Charts/ChartistBar.razor
@@ -1,6 +1,3 @@
-﻿@using Microsoft.JSInterop;
+﻿@inherits Chartist.Blazor.Core.ChartBase
 
-
-
-<div @ref="elem" class="ct-chart ct-golden-section" @attributes="@ExtraAttributes"></div>
-
+<div @ref=elem class="ct-chart ct-golden-section" @attributes=ExtraAttributes></div>
