--- src/Chartist.Blazor/Charts/ChartistEventedBar.razor
@@ -1,3 +1,0 @@
-﻿@inherits Chartist.Blazor.Core.EventedChartBase
-
-<div @ref="elem" class="ct-chart ct-golden-section" @attributes="@ExtraAttributes"></div>
